Removed references to non-existent eigen.i.
this change happens after the merge back to trunk, is is just for sanity 
purposes. I'd been experimenting with swigging the Eigen libraries, and 
this line got left behind on the check in. On the off chance someone 
checks out this ticket I want the ticket to compile first time.