Made a few changes to Footprint and FootprintSet to get rid of bare pointers which are not exception safe and were causing memory leaks.
In Footprint:
Changed _spans and _peaks to std::vector, rather than std::vector&
(which were both initialized as *new std::vector<...> and then the address
of the reference was deleted in the destructor).  These didn't require any
such functionality, so leave them as normal vector's and let std::vector
do the memory management itself.  

By the way, it doesn't seem like _peaks is being used.  The only use is 
in a line that is commented out.  Should this member be removed?

In FootprintSet:
Changed _footprints and _region to be boost::shared_ptr<> rather than 
a reference (which again was initialized with *new, and only _footprints
was being deleted).  I think these could probably also switch to being
stored by value, but that would technically change some of the functionality
of the class -- specifically shallow copies.  I believe storing these
values as shared_ptr's keeps the shallow copy behavior, but now without
the memory leaks.  (It is also more exception safe now.)  However, it 
should be considered whether this is the desired behavior for this class
or if shallow copies were an oversight.


Furthermore, I checked that no other pieces of code in afw contain the 
string "*new ", so I believe there aren't any other similar 
memory leaks lurking in other parts of afw at least.  I didn't grep
through all the other modules, though.